The player can no longer get stuck on the top of a Line.
It was caused when calculatin the distance between the player and the
Line, where we divided the width and height of both objects by 2. This
meant that objects with an odd width of height would be reported as
being a pixel closer than they realy are.

Now doubles are used when calculating the distance, and the result is
then rounded to an int.