Allow SDL_Scancodes to be queried directly
The Keyboard class asks for a char string to check if a key was down.

This is handled be using the function SDL_GetScancodeFromName. This
function searches through an array, comparing the input string until it
finds a match, returning the correct scancode.

The overhead of this actually affects the game, slowing down the game to
a point where frame drops are noticeable.

The fix that's just been added is allowing SDL scancodes to be queried,
as well as strings. This way there's no overhead of searching (it's just
returning an array element).